# openapi.json
{"openapi":"3.1.0","info":{"title":"Employee Survey API","version":"0.1.0"},"paths":{"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/surveys":{"get":{"tags":["surveys"],"summary":"List Surveys","operationId":"list_surveys_admin_surveys_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SurveyOut"},"type":"array","title":"Response List Surveys Admin Surveys Get"}}}}}},"post":{"tags":["surveys"],"summary":"Create Survey","operationId":"create_survey_admin_surveys_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SurveyCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SurveyOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/surveys/{sid}/questions":{"get":{"tags":["responses"],"summary":"Get Questions","operationId":"get_questions_surveys__sid__questions_get","parameters":[{"name":"sid","in":"path","required":true,"schema":{"type":"integer","title":"Sid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/surveys/{sid}/responses":{"post":{"tags":["responses"],"summary":"Submit Response","operationId":"submit_response_surveys__sid__responses_post","parameters":[{"name":"sid","in":"path","required":true,"schema":{"type":"integer","title":"Sid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AnswerIn"},"title":"Payload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnswerIn":{"properties":{"question_id":{"type":"integer","title":"Question Id"},"value_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Value Text"},"value_numeric":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Value Numeric"}},"type":"object","required":["question_id"],"title":"AnswerIn"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginIn":{"properties":{"email":{"type":"string","title":"Email"},"role":{"type":"string","title":"Role"}},"type":"object","required":["email","role"],"title":"LoginIn"},"QuestionIn":{"properties":{"qtype":{"type":"string","title":"Qtype"},"text":{"type":"string","title":"Text"}},"type":"object","required":["qtype","text"],"title":"QuestionIn"},"SessionOut":{"properties":{"token":{"type":"string","title":"Token"},"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role"}},"type":"object","required":["token","name","role"],"title":"SessionOut"},"SurveyCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description","default":""},"anonymous":{"type":"boolean","title":"Anonymous","default":true},"questions":{"items":{"$ref":"#/components/schemas/QuestionIn"},"type":"array","title":"Questions","default":[]}},"type":"object","required":["title"],"title":"SurveyCreate"},"SurveyOut":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"}},"type":"object","required":["id","title"],"title":"SurveyOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}